+++ e2e/app/modal/annotation-field-modal.ts
@@ -1,5 +1,4 @@
 import {Page} from 'puppeteer';
-import {savePageToFile, takeScreenshot} from 'utils/save-file-utils'
 import {LinkText} from 'app/text-labels';
 import Button from 'app/element/button';
 import Modal from './modal';
@@ -23,16 +22,9 @@ export default class AnnotationFieldModal extends Modal {
     super(page, xpath);
   }
 
-  async waitForLoad(): Promise<this> {
-    await super.waitForLoad();
-    try {
-      await waitForText(this.page, modalTitle, {xpath: this.getXpath()});
-    } catch (e) {
-      await savePageToFile(this.page);
-      await takeScreenshot(this.page);
-      throw new Error(`AnnotationFieldModal waitForLoad() encountered ${e}`);
-    }
-    return this;
+  async isLoaded(): Promise<boolean> {
+     await waitForText(this.page, modalTitle, {xpath: this.getXpath()});
+     return true;
   }
 
   /**
