+++ e2e/app/modal/edit-delete-annotations-modal.ts
@@ -1,13 +1,10 @@
 import {Page} from 'puppeteer';
-import {savePageToFile, takeScreenshot} from 'utils/save-file-utils'
 import {LinkText} from 'app/text-labels';
 import Modal from './modal';
 import {waitForText} from 'utils/waits-utils';
 import BaseElement from 'app/element/base-element';
 
-
 const modalTitle = 'Edit or Delete Review-Wide Annotation Fields';
-
 
 export default class EditDeleteAnnotationsModal extends Modal {
 
@@ -15,18 +12,10 @@ export default class EditDeleteAnnotationsModal extends Modal {
     super(page, xpath);
   }
 
-  async waitForLoad(): Promise<this> {
-    await super.waitForLoad();
-    try {
-      await waitForText(this.page, modalTitle, {xpath: this.getXpath()});
-    } catch (e) {
-      await savePageToFile(this.page);
-      await takeScreenshot(this.page);
-      throw new Error(`EditOrDeleteAnnotationFieldModal waitForLoad() encountered ${e}`);
-    }
-    return this;
+  async isLoaded(): Promise<boolean> {
+     await waitForText(this.page, modalTitle, {xpath: this.getXpath()});
+     return true;
   }
-
 
    /**
     * @param {string} newName New name.
