+++ e2e/app/page/workspace-edit-page.ts
@@ -1,4 +1,3 @@
-import Modal from 'app/component/modal';
 import Button from 'app/element/button';
 import Checkbox from 'app/element/checkbox';
 import Select from 'app/element/select';
@@ -9,6 +8,7 @@ import {ElementHandle, Page} from 'puppeteer';
 import {waitForDocumentTitle, waitWhileLoading} from 'utils/waits-utils';
 import {buildXPath} from 'app/xpath-builders';
 import {LinkText} from 'app/text-labels';
+import Modal from 'app/modal/modal';
 import WorkspaceBase, {UseFreeCredits} from './workspace-base';
 import {config} from 'resources/workbench-config';
 import BaseElement from 'app/element/base-element';
