+++ e2e/tests/notebook/notebook-owner-actions.spec.ts
@@ -1,5 +1,5 @@
 import DataResourceCard from 'app/component/data-resource-card';
-import NewNotebookModal from 'app/component/new-notebook-modal';
+import NewNotebookModal from 'app/modal/new-notebook-modal';
 import WorkspacesPage from 'app/page/workspaces-page';
 import {LinkText, ResourceCard} from 'app/text-labels';
 import {makeRandomName, makeWorkspaceName} from 'utils/str-utils';
