+++ e2e/tests/workspace/cdr-version-upgrade.spec.ts
@@ -5,7 +5,7 @@ import {config} from 'resources/workbench-config';
 import WorkspaceDataPage from 'app/page/workspace-data-page';
 import WorkspaceBase from 'app/page/workspace-base';
 import WorkspaceEditPage from 'app/page/workspace-edit-page';
-import CdrVersionUpgradeModal from 'app/page/cdr-version-upgrade-modal';
+import CdrVersionUpgradeModal from 'app/modal/cdr-version-upgrade-modal';
 
 describe('Workspace CDR Version Upgrade modal', () => {
     beforeEach(async () => {
